requirements.txt
instagrapi>=2.1.2
Pillow>=10.0.0
requests>=2.28.0
